pyproject.toml
[project]
name = "woodland-chess"
version = "0.1.0"
description = "Club chess analytics MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.44.0",
  "plotly>=5.20.0",
  "pandas>=2.2.0",
  "sqlalchemy>=2.0.29",
  "psycopg[binary]>=3.2.0",
  "python-chess>=1.999",
  "streamlit-plotly-events>=0.0.6",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.8.0",
  "requests>=2.31.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
